+++ src/main/kotlin/klib/extensions/Long.kt
@@ -1,0 +1,22 @@
+/*
+ * This file belongs to the source code of: kLib
+ *
+ * Copyright© 2015-2021 Thomas Obernosterer, ATVG-Studios
+ * Copyright© 2019-2021 all kLib Contributors
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
+ * If a copy of the MPL was not distributed with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+package klib.extensions
+
+/**
+ * Converts a Long into a ULong
+ *
+ * @returns Long as ULong
+ *
+ * @since 6.0.0
+ * @author Thomas Obernosterer
+ */
+val Long.ul: ULong
+    get() = this.toULong()